fixes for macos

+++ Engine2.cpp
@@ -994,7 +994,7 @@ QByteArray Engine2::getTypeName(int arg) const
 
 static QByteArray _toHex(const void *p)
 {
-    return "0x" + QByteArray::number((quint32)p, 16 ); // table, thread, function, userdata
+    return "0x" + QByteArray::number((quint64)p, 16 ); // table, thread, function, userdata
 }
 
 QByteArray Engine2::getValueString(int arg, bool showAddress ) const
